techDebt(fixTrianingSpelling): Rename trianing-link-panel component and update import statements
