+++ src/app/shared/components/training-link-panel/training-link-panel.component.html
